% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_parameter.R
\name{compute_parameter}
\alias{compute_parameter}
\title{Estimate Conditional Expectations and Propensities for Semi-Supervised ATE}
\usage{
compute_parameter(K, Y, A, X, S, W, foldid, R, cf_model)
}
\arguments{
\item{K}{Integer. Number of cross-fitting folds.}

\item{Y}{Numeric vector. Outcome variable (can contain NAs for unlabelled data).}

\item{A}{Numeric vector. Treatment assignment (0 or 1, can contain NAs).}

\item{X}{Matrix or data frame. Covariates used for outcome and propensity estimation.}

\item{S}{Matrix or data frame. Additional covariates only used in imputation models.}

\item{W}{Matrix or data frame. Combination of X and S.}

\item{foldid}{Integer vector. Fold assignments for cross-fitting.}

\item{R}{Binary vector. 1 indicates labelled data (non-missing A and Y), 0 indicates unlabelled.}

\item{cf_model}{Function. Cross-fitting wrapper function, e.g., using Super Learner or another learner.}
}
\value{
A list containing estimated nuisance parameters (each is a vector):
\describe{
\item{pi1.bs}{Estimated propensity score \eqn{P(A = 1 | X)}}
\item{pi0.bs}{Estimated propensity score \eqn{P(A = 0 | X)} = 1 - pi1.bs}
\item{mu1.bs}{Estimated outcome regression \eqn{E[Y | A = 1, X]}}
\item{mu0.bs}{Estimated outcome regression \eqn{E[Y | A = 0, X]}}
\item{cap_pi1.bs}{Estimated imputed propensity score \eqn{P(A = 1 | W)}}
\item{cap_pi0.bs}{Estimated imputed propensity score \eqn{P(A = 0 | W)} = 1 - cap_pi1.bs}
\item{m1.bs}{Estimated imputed outcome regression \eqn{E[Y | A = 1, W]}}
\item{m0.bs}{Estimated imputed outcome regression \eqn{E[Y | A = 0, W]}}
}
}
\description{
Computes nuisance parameters such as conditional expectations and propensity scores
using cross-fitting for both labelled and unlabelled data. This function is central
for ATE estimation under semi-supervised or doubly robust frameworks.
}
