% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example Dataset}
\format{
\subsection{\code{dat}}{

A data frame(N=100):
\describe{
\item{Y}{Observed outcomes}
\item{A}{Treatment indicator (1 for treated, 0 for control)}
\item{X}{Year}
\item{S}{}
\item{pi}{}
\item{mu1}{}
\item{mu0}{}
\item{imp.meas}{}
}
}
}
\usage{
dat
}
\description{
This is an example dataset included in the package.
}
\keyword{datasets}
