% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fun.R
\name{fit.fun}
\alias{fit.fun}
\title{Fit Logistic Regression and Extract Coefficients}
\usage{
fit.fun(formula, data, method, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} (e.g., \code{response ~ predictor1 + predictor2}) describing the model to be fitted.}

\item{data}{Data}

\item{...}{Additional arguments to be passed to \code{glm}, such as \code{data}, \code{weights}, or \code{subset}.}

\item{methos}{Which model to use}
}
\value{
A numeric vector containing the coefficients of the fitted logistic regression model.
}
\description{
This function fits a logistic regression model to the given data using the specified formula
and returns the model's coefficients. It serves as a simplified wrapper for \code{glm} with a binomial family.
}
