% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{main_function}
\alias{main_function}
\title{Title}
\usage{
main_function(Y, A, S, X, K, folds)
}
\arguments{
\item{Y}{Numeric vector. Observed outcomes.}

\item{A}{Numeric vector. Treatment indicator (binary or categorical).}

\item{S}{Matrix. Auxiliary predictor matrix (\eqn{n \times q}).}

\item{X}{Matrix. Predictor matrix (\eqn{n \times p}).}

\item{K}{Integer. The number of folds for cross-validation.}

\item{folds}{Integer vector. Fold assignments for each observation, with values from 1 to \eqn{K}.}
}
\value{
Numeric. The estimated Average Treatment Effect (ATE) using SSL methods.
}
\description{
Title
}
