% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_fun.R
\name{param_fun}
\alias{param_fun}
\title{Parameter grid function}
\usage{
param_fun()
}
\value{
A list, e.g., list(ridge = ..., lambda = ...)
}
\description{
Returns two list of hyperparameters for model tuning.
}
