% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_ate.R
\name{random_ate}
\alias{random_ate}
\title{Title}
\usage{
random_ate(Y, A, mu_a, m_a, pi_1, pi_0, Pi_a)
}
\arguments{
\item{Y}{Numeric vector. Observed outcomes.}

\item{A}{Numeric vector. Treatment indicator (1 for treated, 0 for control).}

\item{mu_a}{Numeric vector. Predicted outcomes under treatment.}

\item{m_a}{Numeric vector. Predicted outcomes from auxiliary models.}

\item{pi_1}{Numeric vector. Propensity scores for the treatment group.}

\item{pi_0}{Numeric vector. Propensity scores for the control group.}

\item{Pi_a}{Numeric vector. Propensity scores from auxiliary models.}
}
\value{
A numeric value representing the randomized ATE (SSL-based estimate).
}
\description{
Title
}
