% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_models.R
\name{refine_models}
\alias{refine_models}
\title{Title}
\usage{
refine_models(K, folds, X, A, Y, X_matrix_raw, Y_matrix_raw)
}
\arguments{
\item{K}{Integer. The number of folds for cross-validation.}

\item{folds}{Integer vector. Fold assignments for each observation, with values from 1 to \eqn{K}.}

\item{X}{Matrix. Predictor matrix (\eqn{n \times p}), where \eqn{n} is the number of observations
and \eqn{p} is the number of predictors.}

\item{A}{Numeric vector. Treatment or intervention indicator (binary or categorical).}

\item{Y}{Numeric vector. Observed outcomes.}

\item{X_matrix_raw}{Matrix. Raw predictor matrix to be refined (\eqn{n \times p}).}

\item{Y_matrix_raw}{Matrix. Raw outcome matrix to be refined (\eqn{n \times m}).}
}
\value{
A list containing:
\describe{
\item{X_matrix}{Refined predictor matrix (\eqn{n \times p}).}
\item{Y_matrix}{Refined outcome matrix (\eqn{n \times m}).}
}
}
\description{
Title
}
