% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_models}
\alias{train_models}
\title{Title}
\usage{
train_models(K, folds, X, A)
}
\arguments{
\item{K}{Integer. The number of folds for cross-validation.}

\item{folds}{Integer vector. Fold assignments for each observation, with values from 1 to \eqn{K}.}

\item{X}{Matrix. Predictor matrix (\eqn{n \times p}), where \eqn{n} is the number of observations
and \eqn{p} is the number of predictors.}

\item{A}{Numeric vector. Treatment or intervention indicator (binary or categorical).}
}
\value{
A list containing:
\describe{
\item{X_matrix_raw}{Raw predictor matrix (\eqn{n \times p}) generated during model training.}
\item{Y_matrix_raw}{Raw outcome matrix (\eqn{n \times m}) generated during model training.}
}
}
\description{
Title
}
