% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_parameter.R
\name{compute_parameter}
\alias{compute_parameter}
\title{Title}
\usage{
compute_parameter(
  K,
  folds,
  X,
  S,
  X_matrix,
  Y_matrix,
  parameter_list_W,
  parameter_list_Y
)
}
\arguments{
\item{K}{Integer. The number of folds for cross-validation.}

\item{folds}{Integer vector. Fold assignments for each observation.}

\item{X}{Matrix. Predictor matrix (\eqn{n \times p}) used for model training.}

\item{S}{Matrix. Auxiliary predictor matrix (\eqn{n \times q}) used for training.}

\item{X_matrix}{Matrix. Refined predictor matrix (\eqn{n \times p}) for storing intermediate results.}

\item{Y_matrix}{Matrix. Refined response matrix (\eqn{n \times m}) for storing intermediate results.}

\item{parameter_list_W}{List. A list of parameters associated with the combined design matrix \eqn{W}.}

\item{parameter_list_Y}{List. A list of parameters associated with the response matrix \eqn{Y}.}
}
\value{
A list containing:
\describe{
\item{pi_1}{Numeric. Computed propensity score for the treatment group (\eqn{\pi_1}).}
\item{pi_0}{Numeric. Computed propensity score for the control group (\eqn{\pi_0}).}
\item{mu_a}{Numeric. Predicted outcome under treatment.}
\item{Pi_a}{Numeric. Predicted propensity score from auxiliary models.}
\item{m_a}{Numeric. Auxiliary outcome predictions.}
}
}
\description{
Title
}
