% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Title}
\usage{
cross_validation(K, folds, X, S, A)
}
\arguments{
\item{K}{Integer. The number of folds for cross-validation.}

\item{folds}{Integer vector. Fold assignments for each observation, with values from 1 to \eqn{K}.}

\item{X}{Matrix. Predictor matrix (\eqn{n \times p}), where \eqn{n} is the number of observations
and \eqn{p} is the number of predictors.}

\item{S}{Matrix. Auxiliary predictor matrix (\eqn{n \times q}) used in model training.}

\item{A}{Numeric vector. Treatment or intervention indicator (binary or categorical).}
}
\value{
A list containing:
\describe{
\item{parameter_list_W}{List of parameters associated with the combined predictor matrix \eqn{W}.}
\item{parameter_list_Y}{List of parameters associated with the outcome matrix \eqn{Y}.}
}
}
\description{
Title
}
